UI update, permissions fix, minor bug fixes

+++ server/routes/quizQuestions.js
@@ -5,33 +5,34 @@ const router = express.Router()
 
 router.post('/', async (req, res) => {
     try {
-        const {questionText,options,quizId,order,answerType} = req.body
+        const { questionText, options, quizId, order, answerType } = req.body
         const newQuizQuestion = await QuizQuestionService.createQuizQuestion({
             questionText,
             options,
             quizId,
             order,
-            answerType
+            answerType,
+            userId: req.userId
         })
-        res.json({ data: newQuizQuestion, message: "Question created successfully!" , status: "success" })
+        res.json({ data: newQuizQuestion, message: "Question created successfully!", status: "success" })
     } catch (err) {
-        res.json({ data: null, message: err.message , status: "error" })
+        res.json({ data: null, message: err.message, status: "error" })
     }
 })
 
 router.delete('/:id', async (req, res) => {
     try {
-        const {id} = req.params
-        const {order} = req.query
+        const { id } = req.params
+        const { order } = req.query
         if (order !== undefined) {
-            await QuizQuestionService.deleteQuizQuestion({quizId: id, order})
+            await QuizQuestionService.deleteQuizQuestion({ quizId: id, order: Number(order), userId: req.userId })
             res.json({ data: null, message: "Question deleted successfully!", status: "success" })
         } else {
-            await QuizQuestionService.deleteQuizQuestions({quizId: id})
+            await QuizQuestionService.deleteQuizQuestions({ quizId: id, userId: req.userId })
             res.json({ data: null, message: "Questions deleted successfully!", status: "success" })
         }
     } catch (err) {
-        res.json({ data: null, message: err.message , status: "error" })
+        res.json({ data: null, message: err.message, status: "error" })
     }
 })
